requirements.txt
asgiref==3.2.10
automium==0.2.6
automium-web==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.7
djangorestframework==3.11.0
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
MarkupSafe==1.1.1
psycopg2==2.8.5
python-environ==0.4.54
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
